fix: fix bugs and dead code in pico apss (#15)
* Fix bugs, dead code, and inconsistencies found in C firmware review

Bugs fixed:
- lidar.c: lidar_reset() changed from bool (with no return) to void
- imu.cpp: IMU now detects sensor timeout (IMU_EVENT_TIMEOUT_MS) and
  re-initializes automatically, closing the gap where is_initialized
  was never cleared on mid-operation failures
- imu.cpp: "calibrated" field changed from KV_STR "True"/"False" to
  KV_BOOL for protocol consistency
- motor.c: skip stepper enable/disable GPIO toggling when nsteps == 0
- rfswitch.c: removed redundant ternary inside null check

Dead code removed:
- main.c: unused uid_str computation and pico/unique_id.h include
- lidar.c: unused sf30d_init() function, stale comments, unused cJSON include
- imu.cpp: commented-out calibration block replaced with explanation

Cleanup:
- temp_simple.h: added DS18B20_CONVERSION_TIME_MS constant (750ms
  hardware limitation) with documentation
- temp_simple.h: documented that get_conversion_time returns absolute
  time (ms since boot), not a duration
- imu.h: fixed parameter name mismatch (index -> app_id)
- imu.cpp: documented zero-initialized sensor_data behavior

Updated Python emulator and tests to match firmware changes.
Added 4 new tests covering the fixed behaviors.


* Add integration tests for status field types and calibration round-trip

The existing integration tests only checked that field *names* existed
in status messages but never verified *types*. This meant a firmware
change like switching "calibrated" from string to boolean would pass
all tests while silently breaking host code.

New tests:
- IMU: verify calibrated is bool (not string) over serial pipeline
- IMU: calibrate() round-trip through PicoIMU -> serial -> emulator
- IMU: all status field types (bool, int, float, str)
- Motor: all status field types
- Peltier: all status field types (bool, numeric, str)

These type-checking tests act as a contract between firmware and host:
if either side changes a field's type, the test fails.


* Add status type tests for TempMon, Lidar, and RFSwitch

Completes the type-checking integration test coverage across all six
device types. Every emulator's status fields now have their types
verified through the full serial pipeline.


* Remove redundant tests and fix calibrated field type (str→bool)

Update tests to match KV_BOOL change for the IMU calibrated field,
and remove ~24 tests from test_emulators.py that were duplicated in
test_protocol_conformance.py across multiple commits. Keeps clean
separation: protocol_conformance for C firmware contract, emulators
for behavioral/convergence tests, integration for serial pipeline.


* Fix conversion_time return type: float→uint32_t to match storage

The getter returned float while the underlying field is uint32_t,
causing silent precision loss after ~4.6h uptime. Use uint32_t
end-to-end and cast to double only at the JSON serialization boundary.


* Add missing newline at end of lidar.h


---------

+++ src/tempmon.c
@@ -30,11 +30,11 @@ void tempmon_server(uint8_t app_id, const char *json_str) {
 }
 
 void tempmon_status(uint8_t app_id) {
-    const float time1 = temp_sensor_get_conversion_time(&sensor1);
-    const float time2 = temp_sensor_get_conversion_time(&sensor2);
+    const uint32_t time1 = temp_sensor_get_conversion_time(&sensor1);
+    const uint32_t time2 = temp_sensor_get_conversion_time(&sensor2);
     const float temp1 = temp_sensor_get_temp(&sensor1);
     const float temp2 = temp_sensor_get_temp(&sensor2);
-    
+
     const char *status1 = temp_sensor_has_error(&sensor1) ? "error" : "update";
     const char *status2 = temp_sensor_has_error(&sensor2) ? "error" : "update";
 
@@ -43,10 +43,10 @@ void tempmon_status(uint8_t app_id) {
         KV_INT, "app_id", app_id,
         KV_STR, "A_status", status1,
         KV_FLOAT, "A_temp", temp1,
-        KV_FLOAT, "A_timestamp", time1,
+        KV_FLOAT, "A_timestamp", (double)time1,
         KV_STR, "B_status", status2,
         KV_FLOAT, "B_temp", temp2,
-        KV_FLOAT, "B_timestamp", time2
+        KV_FLOAT, "B_timestamp", (double)time2
     );
 }
